requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
amqp==5.3.1
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.3.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
croniter==6.0.0
cryptography==41.0.7
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonpath-ng==1.7.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
lxml==6.0.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
ply==3.11
prompt_toolkit==3.0.51
pycparser==2.22
Pygments==2.19.2
PyMySQL==1.1.0
PyPDF2==3.0.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
reportlab==4.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3>=2.5.0
vine==5.1.0
wcwidth==0.2.13
langchain-openai==1.1.6
django-simpleui>=2025.6.24

# UI自动化测试依赖包
playwright>=1.40.0
selenium==4.15.2
webdriver-manager==4.0.1
browser-use==0.10.1

# JWT 安全机制
djangorestframework-simplejwt==5.5.1